HACK: kernelbase: Do not flash the winedbg console window.
Depending on what the crashing software does with the handles / own
console we may end up creating a new console window for windbg.

Since Proton defaults to starting windbg in a non-interactive mode
without displaying the crash dialog and redirects the log to unix's
stderr we want to avoid confusing users.

CW-Bug-Id: #21872