server: Cleanup message queue on thread exit.
When threads are killed they don't get a chance to close their message
queue handle which is kept in user_thread_info->server_queue, and the
message queues are kept alive on the server side.

In this case, msg_queue_destroy and cleanup_results are never called,
leaving message results referencing the stale queue in other thread
message queues.

Later, when the messages from the other threads are cleaned up this
causes invalid shared memory access when we try to update the queue
bits from the message results.

This doesn't cause much trouble upstream because it only leaks queue
objects, which are always valid in memory, but it breaks with Proton
shared memory patches, and the recent unmap fixes, as the queue bits
are now in unmapped memory.

This is probably a hack because it should probably instead close the
handle on behalf of the dying thread, but this doesn't play well with
process handle caches we use for fsync / esync. The handle should be
closed on process exit anyway.