jscript: Implement proxy objects along with a private interface.
Native mshtml objects act like JS objects in IE9+ mode (the JS engine is
implemented in mshtml). This implements a sort of proxy JS objects for
mshtml objects that export a new private (internal to Wine) interface,
to achieve the same behavior. Another callback interface on jscript side
is exposed to mshtml implicitly, which mshtml uses to callback or notify
jscript of changes or things it needs (such as being unlinked from mshtml).

The props that are forwarded to the mshtml side are stored as either proxy
functions or accessors, or a new prop type (for custom values), which is
only available for such proxy JS objects, and use the internal interface to
forward them to mshtml. Some of this interface is close to IDispatch, but is
separate for two reasons: it has to skip asking jscript again for the prop
(unlike when it's accessed by non-JS code for example), and accessors may
potentially have a different "this" pointer as well passed to them.

Since it's purely internal, the private interfaces are not part of any
typelib, and for simplicity they extend IDispatchEx, to keep one vtbl.

A mshtml object can also be accessed by normal COM code, and it exposes
the typical interfaces while also having the jscript props (as we check in
tests), so we have to forward its calls to jscript, which might forward it
back to mshtml, but only once. We can't expose the jscript proxy object
directly at all. However, even on native, its DISPIDs are different than
the ones in the .idl file, so it's fine to forward them to jscript.

We query external IDispatch objects on demand for the interface. If it's
available, we either create a JS proxy object that is associated with
the mshtml object exposing the interface, or re-acquire another old one
(possibly dangling with refcount zero). It is also associated on the mshtml
side, but not via an actual strong ref, so that we keep a 1:1 mapping between
the JS proxy object and the actual mshtml object, so it preserves the props,
prototypes, and DISPIDs, but without creating a cyclic ref.

As a consequence of this, we don't immediately free the jsdisp when its
refcount reaches zero. Instead, we release the mshtml dispatch ref we hold,
and wait for mshtml to callback and unlink us. It does that when it no
longer needs it, such as when the mshtml dispatch was actually destroyed
(no other refs to it). We can then free the jsdisp if it's dangling, or
disassociate it from the proxy otherwise.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>