mshtml: Implement the private jscript proxy interface.
This implements the internal private interface for mshtml objects, so they
can be used as JS proxy objects from jscript.

Note that the association from these objects back to the JS proxy objects
does not hold a ref to them, but only a weak ref, to prevent circular
references. However, because the associated jsdisp remains alive even if
it's dangling with a refcount of 0, the pointers are always valid until
Unlinked is called on them via the callback, which invalidates them.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>