win32u: Send EVENT_SYSTEM_FOREGROUND in another case.
CW-Bug-Id: #23779

EVENT_SYSTEM_FOREGROUND is sent by set_active_window, but in some
cases the active window is set before the foreground window,
meaning the foreground window hasn't changed yet at the time of
the event.

This is most likely wrong, as it leads to duplicate
EVENT_SYSTEM_FOREGROUND events, but solving it correctly is a
thorny problem that will require tests.