winex11: Use the current state when deciding how to reply to WM_TAKE_FOCUS.
The Win32 state might not have been updated with the latest X11 state,
and we want to take future updates into account when deciding to accept
focus or not.

At this point, and because we ignore WM_TAKE_FOCUS when a state change
is pending, all the states are the same, we use current_state as the
logic is to accept focus or not according to our current understanding
of the X state.