winex11: Avoid creating windows with override-redirect flag set.
This confuses mutter, and it will later get confused in various ways if
decorations are later added or removed on the window, even if the flag
has been unset. Doing it the other way, changing the flag later on, when
needed works fine.

This manifests with spurious IconicState WM_STATE change when mutter
decides to try managing the window, but it also makes it randomly lose
focus or even fail to map the window back on screen.