winex11: Give focus to the expected window when unampping with focus.
SetForegroundWindow calls aren't reflected to the X11 side, and we're
sometimes becoming inconsistent between the X focused window and the
Win32 foreground window.

When a window gets unmapped, another one may receive WM_TAKE_FOCUS, and
in the next change we want to rely on the foreground window activation
time to make a decision on whether the foreground window is temporarily
unmapping itself. It needs to be the same as the window with X input
focus, or we won't know which window activation time to check.