winevulkan: Make device memory wrapper available in callbacks.
With VK_EXT_device_address_binding_report we can get debug_util
callbacks used to track memory bindings. Since it's the host's
implementation that starts the callback we have to be sure that we have
a way of converting it to the client side's variant before it's added to
the handle map - i.e. we don't know the host handle at that time yet.

Link: https://gitlab.winehq.org/wine/wine/-/merge_requests/5658