HACK: mfplat: Use the MP4 bytestream handler as a fallback.
Instead of falling back directly to GStreamer media source, giving a
chance to winedmo.

This fixes Twisted Sails video playback, as it plays a MP4 file with
a ".resource" extension, no MIME type, and doesn't set the
MF_RESOLUTION_CONTENT_DOES_NOT_HAVE_TO_MATCH_EXTENSION_OR_MIME_TYPE
flag on resolver creation.

This is a hack as it probably is something wrong with the resolver
creation, which should likely find the MP4 bytestream handler on its
own, even if that flag isn't set, but it will also make sure the MF
pipelines are used all the time as default, similar to Proton 9 NMS.

CW-Bug-Id: #24673