HACK: winebus: Create dedicated threads to write evdev haptics output reports.
Some controllers block on ioctl and write calls, and this may propagate
up to the frontend through various blocking calls, including when the
frontend API is using async I/O because of how Wine implements async I/O
for device objects.

This makes sure we fix it once and for all for the Steam Input virtual
controller at least.

CW-Bug-Id: #24594