win32u: Don't create a window surface if the window has a pixel format.
If the GL surface is offscreen, winex11 enable_direct_drawing will not
catch it and will create a window surface nonetheless. Things are being
moved to win32u, so it's probably better to do the check here.

We still allow window surfaces to exist with VK for now, so only enable
this for GL client surfaces.

CW-Bug-Id: #24946
CW-Bug-Id: #24955