win32u: Don't use the current mode in the registry if it's a detached mode.
Fix The Witcher 3 (292030) shows a black screen after the screen is turned off and then on. This is
due to a detached mode is written as the current mode in the registry and then the detached mode
gets used as the current mode when adding fake modes when handling subsequent display change events.

(cherry picked from commit 47210da284de1b001772650e6fd5489715fc0a4a)

CW-Bug-Id: #24948