mshtml: Make sure manually created document dispex info is initialized in IE9+ modes.
Otherwise, the prototype and all the related info may never be initialized,
as the script can simply access properties of the object without interacting
with the object's methods and so ensure_real_info never ends up called.

Note that we can't do this from within create_document_node, as that will
break if it's a document created from update_window_doc (the window's doc
is not set at that point, since it's being created, and ensure_real_info
needs it).

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>