jscript: Add initial implementation of Typed Arrays.
Typed Arrays have special properties (even according to the spec); they
override all positive indices, and operations on them must be special
cased. Presumably (on native) this is done for performance reasons, as they
are used as actual data arrays which are expected to perform well with many
elements (e.g. using them to store and manipulate an image). The spec is
even stricter, where any number is intercepted, but for us it's simpler
since native doesn't follow the spec here.

Note that even the behavior is special; defining an indexed prop on a typed
array for indices out of bounds of the array doesn't do anything, although it
does throw if flags don't match (since it's not configurable). Redefining
an index within bounds throws if flags don't match, but otherwise acts
like a normal setter (coerces it into the underlying data type), which is
NOT how defineProperty usually works. delete always returns false, even
for out of bounds indices, despite them not being props (hasOwnProperty
returns false), which is another special case that, thankfully, actually
makes our design simpler.

The prototype is also not consulted for any positive indices, but is for
negative (unlike the spec says, native differs here), i.e. the positive
indices behavior is completely overridden, even for those out of bounds.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>