winegstreamer: HACK: Always preserve theora timestamps.
The theora decoder can receive data from FFmpeg that includes multiple
frames, but the decoder will only preserve the metadata for the
first frame that is produced. This results in invalid timestamps (zeros) for
the subsequent frames.

This hack ensures that the timestamps provided by the theora decoder
are always used.

The theora decoder is only used with transcoded video.

This fixes a regression in Darksiders Warmastered Edition (462780) introduced
when cherry-picking an upstream fix for the following bugs:

CW-Bug-Id: #23810
CW-Bug-Id: #24708