winex11: Avoid requesting CWStackMode alone with managed windows.
Some window manager won't send a ConfigureNotify event if the stack
mode isn't actually changed, and we cannot predict whether it will.

We would be waiting forever for an event and delay other requests when
we can just request the above state the next time the window needs to
be reconfigured.

CW-Bug-Id: #25059