win32u: Limit the work area to the monitor rect.
When emulating a non-native aspect ratio resolution, calling map_monitor_rect() to map the physical
work area to the emulate work area might produce a rectangle larger than the emulated monitor
rectangle. For example, a physical work area (0,0,3840,2096) gets mapped to (0,-180,2560,1217) when
the emulated resolution is 2560x1080. In this case, we should limit the work area to the emulated
monitor rectangle.

CW-Bug-Id: #25620